backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
google-generativeai==0.3.2
psycopg[binary]==3.2.9
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
python-multipart==0.0.6
websockets==12.0